Adding Pyro-based RPC at NM level
This new option has been added after ZeroRPC has been showing some problems
that still need to be further investigated.

We are using a single Pyro daemon for the whole Node Manager, and using it in
multiplex mode to avoid resource exhaustion on the server side (although it
doesn't support I/O + coroutines like ZeroRPC does). Pyro had a bug when
shutting down these multiplex servers, but version 4.47 includes a fix we
submitted to fix this.

Subclasses have also been added on the composite manager side to account for
the different RPC mechanisms.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>