Creating and using ZMQ sockets in the same thread
Before we were creating sockets in the main thread and then using them in a
separate thread. This was *apparently* causing some sporadic problems ("socket
operation in non-socket object" errors), specially in python 3, although it's
not 100% sure that was the cause. It's anyway recommended not to share sockets
across multiple threads because they are not thread-safe. These sporadic
problems don't appear anymore after this change, and although that doesn't
necessarily mean they're gone for good I'm fairly confident this was the cause.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>