Use compression for appending graphs
The underlying support was already there, but for some reason we had forgotten
to activate it for the append_graph call. This should allow us to append graphs
that are originally up to 4GB or more, although in the future we will move to
an iterative JSON parsin style instead.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>