ZeroRPC client usage simplified
Instead of having a set of replies to attend for we simply register a callback
on each of the RPC calls that will eventually return the result to the caller
via the return queue. This simplifies our code as we require less objects and
one less greenlet. Performance seems to have improved as well.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>