Streaming json-writing/compression/uncompression
This lets the HTTP clients use less memory to send big graphs and imposes no
limit on the request sizes performed on the server. The server side has already
a streaming consumer for the compressed stream but the JSON content is still
being loaded all at once. Initial tests using ijson were performed and although
it works (therefore reducing the memory footprint) the performance is very
poor, at least in Linux, even when using the CFFI backend which is supposed to
be the fastest.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>