Removed multithreaded access to zmq SUB socket
This was the aparent cause of the unit test failures under python 3. The SUB
socket is now local to the function where it's created (on its on thread), and
subscription information is passed through a shared Queue. Unit tests don't
crash anymore on us, so hopefully the problem is gone for good.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>