Removing templates support on NM
This is an old, unnecessary feature that I added a long time ago without a
proper requirement behind it, thinking that it would be useful some day. So far
we haven't used it, and it only contributes to being a maintenance burden and a
cause for low code coverage during our tests, so I'm removing it for good.

I also took the opportunity to finally remove the NGAS-based test we had for
our NM. It was the only user of the template support on the NM, but relied on
an external NGAS installation anyway.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>