DATA-203: adding git commit to version information
This code is heavily inspired on how numpy achieves the same, but with less
logic (since we don't make source distributions yet, and therefore don't handle
the case of git not being there).

In the future we might want to take those cases into account, but for the time
being this looks like a good solution.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>