Changing loop order in other graph_loader methods
We had changed the looping order on the removeUnmetRelationships method, which
gave us some speed up when traversing the graph and all its drop specs. The
same change has now been applied to the rest of the methods, allowing for
faster execution times in general.

Now methods use mostly the module-level linkType:property maps in the opposite
order, except for one that still benefits on the original mapping order. Thus
the maps now contain the keys and values in both directions; because they have
a one-to-one relationship there is no risk of overlapping.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>