Remove dependencies on example_client, 1st try
Now that all the functionality from the example_client (and more) is present on
the tool module we should be able to port the start_dfms_cluster module to use
the tool module, and eventually remove the example_client one. This is the
first iteration, so there probably are still some bugs here and there. We need
to commit and push though because this has to be tested in Pawsey and this is
the simplest way of doing it.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>