Removing the example_client module
This was actually one of the main points (if not THE one) that motivated the
work on this branch. The functionality that previously resided in the
example_client module now fully lives in the tool module, and since we've
removed any usage of example_client we can safely remove it.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>