Support LG IDs in BashShellApp I/O references.
When a BashShellApp component references inputs/outputs in the Logical Graph it
should do so using the IDs of the Logical Graph Nodes it referes to (which, at
the moment, are negative integers). These IDs should be translated into OIDs
when the LG is translated into a PGT.

This logic was missing as is now available through this commit. I'm not sure
this is the best place to put it though (probably not, as it's very specific
and lives in a very generic code path) but it will make the cut for the time
being.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>