Improvements on the MPIApp
Now the MPIApp class handles two modes of operation: one is calling the MPI
target application directly, in which case it barriers on the
inter-communicator (and thus expects the target application to do the same).
The second mode of operation is what we previously supported, although now we
know it's a bit tricky to get it working correctly.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>