Initial support for daliuge apps written in C
This initial support is in the form of a new BarrierAppDROP called DynlibApp,
which loads a dynamically linked library at initialization time, calls an
initialization routine on the library, and finally a run method on the library.
During initialiation, DynlibApp prepares a representation of itself to be
passed down to the methods invoked in the library. On the C side of things we
include a header file with the definitions of the structures representing the
python application (including callbacks to the read and write methods of the
application's inputs and outputs) and the definitions of the methods that
should be implemented by the application.

The next step is to brake down what is currently implemented in DynlibApp into
more fine-grained logic bits, so we can write a more generic AppDrop version
that supports streaming inputs as well.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>