Improving support for dynlibs
Initial (not yet fully working, but at least designed) support has been
included for more generic streaming C applications loaded from a dynamic linked
library. Speed seems fine, but the dynamic creation of the outputs needs to be
refined (it works fine with 1, not with 2 or more).

The code now is also better documented, specially the C header file.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>