Remove type from dropspect after extraction
The 'type' valye from the dropspec is only needed at this point, and therefore
it should be removed from the dictionary after usage. Failing to do so (as we
were doing until now) means that it will make its way down to the drop class'
initialization parameters, which users can erroneously use.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>