The second python setup.py install was needed
The comment above actually explained it, and was exactly why the Travis builds
were failing. This would not be necessary if the current working directory was
not included in python's sys.path (because then the installed dfms would be
picked up), but it is. This also makes the whole installation a bit pointless,
but again, we need to generate that version file.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>