DEBUG logging to debug a problem in Tianhe-2
A DropProxy is not being attached as an input of an application, which
is making the whole graph fail. It's not immediately obvious from the
code and the current logs what is going on, and the only way I see this
solved is by adding yet more logs (for the time being).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>