Potential fix for spurious dynlib problem
I'm not expert enough in the ctypes library to know exactly why this was
wrong (although to an expert eye it will be probably obvious...), but
this change is fixing a problem we were experiencing with a particular
graph (which we couldn't reproduce with unit tests). The problem itself
was an invalid read, as diagnosed by valgrind.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>