New dfms.apps.simple module and tests
These are the applications that we usually take from
test.graphsRepository. That was not a proper location for them, given
that we use them so much.

Also, the "copy" from "SleepAndCopy" has been factored out, so now we
have a sleeping app, a copying app, and a sleep *and* copy app (while in
the version under the test.graphsRepository module we lacked the
copy-only app).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>