New drop.name attribute
The default value is an empty string, but its value is taken from the
"nm" keyword given to each Drop at initialization time (which, in turn,
has its origin in the current Logical Graph Editor).

This name attribute is then copied into the C structures describing
input and output that are passed down to our dynamic library
applications. This way, applications should be able to more easily
distinguish between different kind of inputs/outputs in the case there
are many of them.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>