DynlibApp calling i.read/o.write directly
The previous option of holding a reference to a read/write method of the
respective object seems to be the cause of some problems we were
experiencing in multi-node deployments. By getting the read and write
attribute first and invoking it later we produce a call to __getattr__
instead of __call__, which an important distinction made by the
DropProxy class at the moment (but that we could probably work around).
By doing always a call to i.read() or o.write() we effectively call
__cal__ on the DropProxy, which is what we need to do.

This problem hasn't been picked up before because we had never actually
tried to run an DynlibApp against remote Drops (we were always using
local Drops). I will add next a unit test that mimics this behavior.
This should both confirm that this is the cause of the problem, and
ensure it doesn't happen again.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>