Testing dynlib apps across Node Managers
We experienced these problems in a multi-node scenario, and only then I
realised that we had not tested this properly. Having this test should
ensure we don't run into this issue anymore.

The two new tests check that both remote inputs and remote outputs work
as expected. The test is simple, but should alert us if there is
something wrong in the future.