Moving to relative imports
Instead of having absolute imports for our own modules, we now switch to
using relative imports. This means that if in the future we decide to
rename a package, we won't need to modify all the files inside that
package anymore.

It is true that we are kind of defeating that very same purpose in this
commit by changing so many files, but this work is being done prior to
actually changing the root package name from dfms to dlg, and therefore
changing all these import statements was inevitable. If we were going to
do all these changes, we better do them once only.