Changing references to dfms package in the code
In some parts of the code the dfms package was hardcoded (i.e., it was
present in some strings). This commit changes all those strings and make
them point to the new dlg package.

We also acknowledge that changing all existing graphs would be way too
much to ask from users, and therefore allow to have apps references
using the old package name (which are internally changed to the new
package name before being used).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>