Correct method association
While playing with some unit tests I noticed that the PRODUCER
relationship had the wrong method associated for linking; the problemw
as that when we were saying A is PRODUCER of B, the code was eventually
calling A.addConsumer(B) instead of calling A.addOutput(B).

The Logical Graph Editor apparently doesn't generate producers and
consumers list in its internal representation (instead it always
generates inputs and outputs list, apparently), and therefore we would
have never stumbled with this problem unless manually testing like I was
doing now.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>