Dynlib outputs in remote Node Managers working
The previous implementation was causing RPCs when the local DropProxy
was being added (because DynlibApp was calling o.oid and o.name). This
race-conditioned with the remote Node Manager hosting that output drop,
who by that time might have or not created the drop. In the case of our
tests we actually deploy each Node Manager sequentially, so the problem
was even easier to reproduce (test comming in the next commit).

Reading the o.oid and o.name was being do to create the C structure
corresponding to that output, so the C library could see it. We don't
actually need to do this at output addition time, and can defer it for
later, which is what we are doing now. I also took the opportunity to
take the _to_c_input and _to_c_output methods out of the DynlibApp
classes and converted them to standalone functions of the module.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>