Renaming package from dlg to daliuge
"dlg" already exists in PyPI, so we can't upload a package with the same
name. "daliuge" instead is not being used, so we can safely take it.

The version string doesn't actually need any change. We are also keeping
still the same directory structure, with the root package still called
"dlg", which was always intentional.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>