Making networkx-based code compatible with 1.X/2.X
Performance seems to be a bit worse now though, which I suspect is
because of the way we use the graph objects (maybe there is some penalty
for maintaining API compatibility?). This affects only the very
long-running tests (which are not run by default), so I'm still
commiting these changes for now.

If in the future if we move to networkx2 only, then we can break
compatibility with 1.X APIs and potentially improve our performance.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>