Be specific about the metis dependency
If we don't specify this, setuptools/pip might still try to get it from
PyPI.

Also, pip requires an extra flag to be passed to interpret the
dependency_links list; otherwise it fails to obtain the particular
version of the package we need. Thus, I'm changing how we use pip in
Travis, and how users should use it in the README file.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>