Moving RPC-related functionality to new rpc module
Now we have also clearly separted the client-related RPC functionality
from the server-related one, producing a final pair of RPCClient and
RPCServer classes that users can use separately if they want.

This separation also means that the definition of the DropProxy class,
which is intrinsically related to any RPC activity, can be also moved
into the rpc module. Moreover, conceptually the DropProxy no longer uses
a NodeManager for its communication, but a more generic RPC client.

The Session class also was receiving the host to which the Node Manager
was binding its different services, without a reason, so this was
removed.

Finally, the zmq_safe utility function was also moved down to the utils
module for better reusage.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>