Adding support for off-process dynlib applications
A new DynlibProcApp spawns a new process where the dynamic library is
actually loaded (instead of being loaded in the invoking process, like
DynlibApp does). Inputs/outputs are exposed via the underlying RPC
mechanism to the new process, which creates an RPC client and the
corresponding DropProxy instances to access them.

In order to be able to transmit the information needed by the
sub-process to create DropProxy instances, we need to find (in the
parent process) what is the correct host/port/session_id triplet to
which the sub-process should connect it. This is done by adding and
_rpc_server attribute to each drop after creation, which will point to
the containing NodeManager (because it is the RPC server serving the
drops). In testing scenarios we launch a stand-alone RPCServer
instance and attach it to the drop in the same way.

Some simplifications have been done that definitely need to be addressed
in the future to make this support more robust, but this is working well
for a first iteration.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>