Coping with new docker API
docker 3 brings in an incompatible change on the container.wait()
method, returning a dictionary with the API result instead of simply the
exit code of the container. With this change we make sure we support
both the old and the new return objects, and therefore both the old and
the new versions of the package.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>