Do not constrain paramiko anymore
This was a "fix" we introduced a long time ago to avoid the cryptography
python packages >= 1.1, which became increasingly more difficult to
compile in old Linux installations due to its dependency on newer
libraries. However, most of the time we now install daliuge via pip,
which in turn means we download pre-compiled versions of most of the
packages, including cryptography, and therefore we don't have this
problem anymore.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>