Enriching PyFuncApp functionality
The previous implementation of PyFuncApp was quite basic, and was still
lacking support for positional arguments, keyword arguments, and some
more complicated function types.

This commits introduces two main improvements. The first is the usage of
"dill", a python package that serializes functions (not only with their
name, but also including its code, closure, globals information, etc.).
Using dill we offer a better support on what which functions can
actually be wrapped using PyFuncApp, including lambdas and inner
functions that use closures (which we previously didn't support).

The second improvement is the initial support for positional and keyword
arguments, thus making the PyFuncApp more widely useful. Users need to
provide extra information on how the different input drops map to the
different parameters. If no input maps to a particular argument, its
default value is used.

A set of unit tests check that these new features work as expected.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>