Improved InMemoryDrop pydata argument
We don't assume anymore that the pydata argument of the InMemoryDrop is
of type bytes, but rather acknowledge that it can be a string type. This
most notably happens when reading the drop description from a
HTTP-transmitted JSON stream, which will yield a string object rather
than a bytes object out of the JSON content.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>