Initial delayed function implementation
The delayed function mimics the dask.delayed function. When invoked, it
creates proxy objects that, when invoked via .compute() build the
Physical Graph corresponding to the code written by the user, schedules
it, submits it to a Node Manager, waits for the result, and returns it
to the user.

At the moment the support is limited to interacting with a single Node
Manager, which means that all computation will happen in a single
computer (although it can be a remote one). We still want to add proper
support for interacting with an Data Island Manager, which means we will
not only submit the graph, but actually partition it and map it to the
individual Node Managers that are currently running on the system.

For easy usage, the delayed function defined in the dask_emulation
module is made visible to users via the top-level dlg package. This
makes user code easily portable, replacing "from dask import delayed"
with "from dlg import delayed" (or similar).

A very rudimentary unit test ensures that the basic functionality is
working as expected.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>