Setting the session on all drops
When drops are created within the context of a session, the session
passes itself down (via the graph_loader functions) to the drops. This
provides a link back to the session from the drop, which in turn allows
users to use some contextual information, like the session ID or other
bits.

There was a previous incarnation of this mechanism earlier, with the
Session passing down *only* the session ID to each drop, and only
*after* drop creation. Since at creation time drops didn't know about
their session, it was somehow implied that they *didn't have* to know.
This new mechanism makes this knowledge more explicit, and thus allows
programmers to rely on the existince of this attribute.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>