LIU-16 Restructuring FileDrop's path calculation
The way we were previously calculating the final path to which a
FileDrop was pointing to were not consistent enough, and also left room
for improvements. In particular, the directory where files were
deposited lived under /tmp, could not be changed at all, and users did
not know its location.

This has been reworked now. If relative paths are given in the
filepath/dirname preferences then they are relative to: a) the current
working directory (which can be changed via the -w command-line option
of the node manager process) and b) the session ID. This makes each file
unique even across graph executions, and easier to find for users. We
still respect absolute paths though, in case users need to point to a
very particular location.

This logic applies also to the DirectoryContainer class. Thus I
separated the (new and existing) common bits into a new PathBasedDrop
class which is inherited by both the DirectoryContainer and FileDrop
classes.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>