Adding structured drop/session to log statements
This is done using a two-part solution: firstly, we need to capture
which session/drop is currently under execution, and secondly we need to
pass this information down to the logging system.

The first part is achieved through a combination of a thread-local
variable and a decorator. For each entity that needs tracking (e.g.,
which drop is currently being executed) we create a thread-local
variable that will contain the entity instance that is currently active
on each thread. To populate this variable we provide a decorator that
can be placed on the object class' method, such that each time a member
method is invoked in an object of that class the decorator will take the
first parameter of that invocation (i.e., "self") and put it in the
thread-local variable. Once the invocation finishes, the thread-local's
value is set to its previous value.  Because tracking sessions and drops
is basically the same operation, we wrote the tracking logic only once
in the dlg.utils module, and reuse it in both the dlg.drop and
dlg.manager.session modules.

The second part of this solution is offering a custom _DlgLogger class
and configuring the logging system to use it. This custom class inspects
the relevant thread-local variables when makeRecord() is called, and
injects the values drop the drop_uid and session_id into the logging
in new fields of the record, therefore making them available in an
structured way to any logging handler/filter that may want to use it.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>