LIU-30 Opting out from doing checksum
For the time being we are still defaulting to *calculating* checksums on
data written to data drops, because default to not doing it would break
too many tests. In the future we can invest some more time to default to
*not* calculating checksums, to adjust the tests, and also maybe even
change the configuration mechanism to be a command-line flag and/or have
a per-drop specification.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>