Avoid TypeError when building exception
We received a report of the close() method being called with a None
descriptor. This should have thrown an exception saying that the
descriptor was invalid, but we could not build the exception in the
first place due to an invalid formatting specifier. This commit makes
sure we handle Nones adequatly, but still doesn't solve the problem that
triggered this to happen, which is still under investigation.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>