Simpler way of closing inputs after invoking app
We were previously creating and handing over a list of drop/descriptor
pairs for each of the input drops of the DynlibApp object so they all
could be closed after the application ran. The resulting code was a bit
cumbersome since it exposed the complexity of having to track the
descriptor of the opened drop along with the drop itself. We are
simplifying this and passing down "closer" lambdas that can be simply
invoked.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>