Simpler OID generation and dropdict reuse in Delayed
We now assume that there is only one thread calling .compute() on an
emulated Delayed object. This makes the code simpler, as we can track
the last number used to generate OIDs statically instead of having to
deal with this as a parameter that gets passed around. We can further
extend this easily to support multithreading by using a thread local
variable in the future.

We are also concentrating now the process of creating an OID, creating
the dictdrop for the Delayed object, assigning the OID, and adding the
objects to the visited set and the graph dictionary in a separate method
so it can be easily reused in the future.

Finally, the dropdict for each Delayed object is not only calculated,
but also cached in a private member. This makes it easier to get the
dropdict in subsequent invocations, and opens the door to simpler
method interfaces in the future.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>