Improving and renaming _DelayedDrop._link
The previous name was slightly confusing (what is linked with what?). Not
it is more clear that we are adding an upstream dependency. The method's
signature is also simpler as we now reuse the dropdict of each Delayed
object.

Thanks to this simpler interface we now don't need to keep track of some
local variables that we were otherwise passing around unnecessarily.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>