An improved _DelayedDrops implementation
The previous implementation of _DelayedDrops was enough when the
drops that made out the list were used individually, but not when the
list itself was used as a single input into a function. This new
implementation injects two new _Delayed drops when the physical graph is
being produced: one application that takes all inputs and produces a
list with all the inputs as elements of the list, and a data drop
representing the list itself, which is finally the input given to the
delayed application receiving the _DelayedDrops object.

A new unit test demonstrates this functionality and makes sure it works
as expected.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>