Correctly dealing with remote exceptions
When a remote exception is thrown, ZeroRPC's async_result will contain
it in its exception field, not its value field. We were previously
blindly returning async_result.value though, therefore never propagating
exceptions to the client side.

This commit rectifies that situation. It takes the async_result and
checks whether it contains a value or an exception. Whatever of the two,
the result is carried out of the thread where the gevent loop is
running, and passed down to the client thread, where it is either thrown
or returned.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>