Centralize handling of queue in subprocess
We use the queue to communicate our parent process how we progress
through the different steps of setting up the dynamic library and
executing it. Centralizing this logic makes it safer to handle the
agreed protocol (put None or an Exception) well in all scenarios, as
well as properly finishing the subprocess when required.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>