Enriching environment of BashShell apps
Now their environment contains two new enviroment variables: DLG_UID
contains the UID of the application drop encapsulating the bash shell
execution, and DLG_SESSION_ID contains the ID of the session to which
that drop belongs, if any. In the future we could add more information
if required.

A new unit test ensured this is working as expected. Also, as a result
of more drop-specific information being passed down to the run_bash
method, a new self._run_bash method has been introduced to avoid code
repetition.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>