NM event listeners handling all events
Because previously we supported only error listeners and not generic
event listeners, the code that added the listeners to the individual
drops constrained the events being forwarded to be only those of type
'status'. This meant that a big number of other events didn't make it to
the user-defined event listeners, which are meant to handle *all* events
and perform filtering by themselves.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>