Communicating logical graph key all the way down
Until now we had no way to connect a particular drop to its node in the
original Logical Graph (if it came from there, that is). We are needing
this nowadays because of the work Vitaliy is doing to "close the loop"
between measuring the execution of a physical graph and putting that
information back into the Logical Graph.

Until now the way oids are constructed by the translator means that one
could use these to extract the Logical Graph key, but it was cumbersome
and required internal knowledge of the formula used to create oids, plus
some parsing. The new way is much easier to use and more explicit about
what this piece of information represents.

The keys from the logical graph nodes need to be passed down from: a)
the logical graph into the physical graph, b) the physical graph into
each drop instance, and c) each drop instance and the events they
generate. This commit includes these three communications, but making
sure drops and events still work for drops that were not created from a
Logical Graph.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>