Correctly serializing default values
We were previously not correctly serializing default values for
functions (in the context of the PyFuncApp and the dask emulation
sitting on top). This problem was only visible when serializing
functions whose default values are not JSON-serializiable, but pick-able
in general. When such a function was found, it couldn't be converted to
JSON, causing a failure on the dlg client when converting the graph into
JSON content for communication to the managers.

New unit tests make sure that this is working as expected.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>