Running delayed tests in different contexts
In order to make sure our delayed function is working properly we now
run our unit tests not only under our delayed implementation, but also
under dask's (if available) and directly (i.e., without delayed). The
mechanism we are using to abstract this is similar to what they did in
the ARL's arlexecute module.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>