Correctly dealing with delayed's nout argument
Whenever the nout argument is received by dask's delayed function, the
returned object can be subscripted, even if nout is 1 or 0. We did not
support this previously (we were returning a single _DataDrop, which
obviously doesn't allow subscripting), and thus we were failing to
correctly imitate dask's functionality in these cases. Moreover, we were
using the _DelayedDrops class to back up this nout argument, which is
errnoeous, as the runtime's result should still be handled regularly
regardless of the advised nout value.

In reality what we needed was yet another class to implement the extra
logic of allowing iteration and indexing on local delayed objects. Both
operations basically generate new delayed objects, which is exactly what
Dask does.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>