Fix RTD API documentation
The autodoc-generated API documentation on RTD was failing to build
because the modules that we import are not actually present in the RTD
virtualenv (and don't seem to get installed when one tells RTD to
install our project). In any case, the solution proposed commonly by RTD
is to mock the modules, which works fine.

This is purely a documentation fix, so let's [skip ci]

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>