Supporting compute over lists of delayed objects
We now support receiving a list/tuple of delayed objects, which is
turned into the corresponding _DelayedDrops object acting as the new
root of the graph. In order to be able to act on user-created lists the
compute method must now reside not only as an attribute of the
DelayedDrop class, but also as a function on the dask_emulation module
(with the _DelayedDrop method calling the top-level function).

We are adding new unit tests that ensure this support is working as
expected across all "delayed" flavours, which also required a bit of
tuning on how we compute dask's delayed objects too.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>