Off-process dynlib apps gracefully cancelled
When cancel() is called on a DynlibProcApp, the underlying process is
terminated after the drop is moved to the CANCELLED state. A unit test
makes sure this happens in a timely fashion.

The timing of things in DynlibProcApp could be improved though, as at
the moment there are a number of possible race conditions that can arise
from the simultaneous start of the underlying process and the transition
to the CANCELLED state (which explains the time.sleep(1) in the unit
test before the call to cancel()).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>