New progress-bar visualization
The progress bar consists on a series of rectangles drawn consecutively
one after another, and representing the number of drops that are on any
given state. We place the finished/completed rectangles first, the
in-progress ones later, then those representing an error, then the
cancelled ones, and finally those representing pending tasks. To keep
the experience consistent we reuse the colors used by the graph nodes to
indicate the states.

For the time being we are using a fixed width of 1000 pixels to draw the
progress bar. In the future we can change to be more responsive, if
necessary.

Also, since this is the simples view of all, and the one taking less
time to load, we're changing the code to use this as the default
visualization.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>